http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@505751 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/commands/WireFormatInfo.cpp
@@ -86,7 +86,7 @@ bool WireFormatInfo::isValid() const {
         return false;
     }
 
-    for( int i = 0; i < magic.size(); i++ ) {
+    for( size_t i = 0; i < magic.size(); i++ ) {
         if( magic[i] != MAGIC[i] ) {
             return false;
         }
